like function added to QuestionAnswerCard

+++ lib/services/cloud_fire.dart
@@ -2,6 +2,7 @@ import 'dart:async';
 
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:cop_belgium_app/models/podcast_info_model.dart';
+import 'package:cop_belgium_app/models/question_answer_model.dart';
 import 'package:cop_belgium_app/models/teaching_clip_model.dart';
 import 'package:cop_belgium_app/models/testimony_model.dart';
 import 'package:cop_belgium_app/models/user_model.dart';
@@ -22,10 +23,10 @@ class CloudFire {
       final hasConnection = await _connectionChecker.checkConnection();
 
       if (hasConnection) {
-        if (user.uid != null && user.email != null) {
+        if (user.id != null && user.email != null) {
           await _firebaseFirestore
               .collection('users')
-              .doc(user.uid)
+              .doc(user.id)
               .set(user.toMap());
         }
       } else {
@@ -58,12 +59,10 @@ class CloudFire {
     }
   }
 
-  Stream<UserModel?>? getUserStream() {
+  Stream<UserModel?>? getUserStream({required String uid}) {
     try {
-      final docSnap = _firebaseFirestore
-          .collection('users')
-          .doc(_firebaseAuth.currentUser?.uid)
-          .snapshots();
+      final docSnap =
+          _firebaseFirestore.collection('users').doc(uid).snapshots();
 
       return docSnap.map((doc) {
         if (doc.data() != null) {
@@ -306,7 +305,7 @@ class CloudFire {
   }
 
   // Testimonies
-  Future<void> createTestimony({required TestimonyModel testimony}) async {
+  Future<bool?> createTestimony({required TestimonyModel testimony}) async {
     try {
       bool hasConnection = await _connectionChecker.checkConnection();
 
@@ -315,6 +314,7 @@ class CloudFire {
               testimony.toMap(),
             );
         docRef.update({'id': docRef.id});
+        return true;
       } else {
         throw ConnectionNotifier.connectionException;
       }
@@ -339,7 +339,7 @@ class CloudFire {
     }
   }
 
-  Future<void> updateTestimony({required TestimonyModel testimony}) async {
+  Future<bool?> updateTestimony({required TestimonyModel testimony}) async {
     try {
       bool hasConnection = await _connectionChecker.checkConnection();
 
@@ -356,6 +356,7 @@ class CloudFire {
             .update({
           'updatedAt': DateTime.now(),
         });
+        return true;
       } else {
         throw ConnectionNotifier.connectionException;
       }
@@ -375,6 +376,70 @@ class CloudFire {
         await _firebaseFirestore
             .collection('testimonies')
             .doc(testimonyModel.id)
+            .delete();
+        return true;
+      } else {
+        throw ConnectionNotifier.connectionException;
+      }
+    } on FirebaseException catch (e) {
+      debugPrint(e.toString());
+      rethrow;
+    } catch (e) {
+      debugPrint(e.toString());
+      rethrow;
+    }
+  }
+
+  // QuestionAnswer
+
+  Stream<List<QuestionAnswerModel>> getQuestionAnswersStream() {
+    try {
+      final qSnap =
+          _firebaseFirestore.collection('question_answers').snapshots();
+
+      return qSnap.map((qSnap) {
+        return qSnap.docs.map((e) {
+          return QuestionAnswerModel.fromMap(map: e.data());
+        }).toList();
+      });
+    } catch (e) {
+      debugPrint(e.toString());
+      rethrow;
+    }
+  }
+
+  Future<bool?> createQuestionAnwsers({
+    required QuestionAnswerModel questionAnswerModel,
+  }) async {
+    try {
+      bool hasConnection = await _connectionChecker.checkConnection();
+
+      if (hasConnection) {
+        final docRef =
+            await _firebaseFirestore.collection('question_answers').add(
+                  questionAnswerModel.toMap(),
+                );
+        docRef.update({'id': docRef.id});
+        return true;
+      } else {
+        throw ConnectionNotifier.connectionException;
+      }
+    } catch (e) {
+      debugPrint(e.toString());
+      rethrow;
+    }
+  }
+
+  Future<bool?> deleteQuestion({
+    required QuestionAnswerModel questionAnswerModel,
+  }) async {
+    try {
+      bool hasConnection = await _connectionChecker.checkConnection();
+
+      if (hasConnection) {
+        await _firebaseFirestore
+            .collection('question_answers')
+            .doc(questionAnswerModel.id)
             .delete();
         return true;
       } else {
